Update build script to remove the build folder from Catch2 library.
Unfortunately, I don't know how to ignore changes in a submodule without
changing the "inner" .gitignore, i.e. the .gitignore in the submodule.
I can't just ignore the directory lib/Catch2/ as well, because we still
want to get the most recent changes from that git-repo, which won't
happen if it is ignored! Hence, I just change the build script. Sry mom.