Allow recursive calls in the ast_printer.
We discussed about how to recursively call an ast_visitor as it is
defined until the last commit. The key idea to recursive calls, since an
ast_visitor in our case is just a lambda, is to basically add a
"recursor" as argument, which should be the function itself. Thankfully
to C++17, this is not painful anymore and can be just `auto recursor` in
the parameter list of a lambda, which is just a syntactic sugar for
templates.