Rename all ast nodes.
This commit probably breaks a lot of stuff that is not pushed.
However, it improves readability and reduces a lot of pain in the long
run.
Since now, we don't have to do
`std::make_unique<rec_wrap_t<error>>(ast_tags::error, old)` anymore,
a simple `ast_tags::error.make_node(old)` suffices.
In turn, the major changes are:

 - Add `make_node` to ast tags, allowing us to not worry anymore about them
   + Eliminate `make_unique`
 - Rename nodes (e.g. `literal` -> `literal_`).
   + Eliminate `rec_wrap_t`